Create README.md
This is the first model printed on RepRap for class. The guidelines were to print something that is no bigger than 4cm in either direction. 
My model is an exploration of creating something similar to an architectural model, dividing space. I was also curious of how delicate details will 
look in such a small model, as you can see the columbs that were 1 or 2mm did not survive long.